fix: [project-detail] 프론트 내 댓글 작성,수정,삭제 기능 보완
1) 프론트 js파일 분리
- 기존 detail.js에 합쳐져 있던 코멘트 기능을 별도 파일로 분리

2) 프론트 페이지에서 로그인 후 댓글 작성,수정,삭제 기능
- 기존 Auth 없던 부분 추가
- 본인이 작성한 댓글만 수정 삭제가 가능하게끔 백엔드에서만 구현됨

3) 백엔드 라우터 파일 분리했던 것을 comments.router.js로 합침
